clean up builtin function application encoding
